Move `ProjectCard` into components.
- It will be used in `/mine` and `/` so make it shared.
