+++ frontend/src/components/ProjectCard.module.scss
